refactor: extract loglines further
